Removed dir.h cuz io.h does the thing. Renamed files, functions and structs related to state search merge (since its rather easy to change between A*/BFS/DFS by changing priority in the state queue)

+++ misc.c
@@ -633,7 +633,7 @@ LOAD_LIST init_load_list(){
 // calculates distance of two 3D points
 int point_distance_3D(short int x1, short int x2, short int y1, short int y2, short int z1, short int z2)
 {
-    return sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2) + pow(z1 - z2, 2));
+    return (int) sqrt(pow(x1 - x2, 2) + pow(y1 - y2, 2) + pow(z1 - z2, 2));
 }
 
 // misc convertion thing
